Refactored points table
