+++ apps/scoutgame/components/claim/components/PointsTable/ClaimedPointsTable.tsx
