+++ apps/scoutgame/lib/points/getPointsWithEvents.ts
@@ -1,5 +1,5 @@
 import { prisma } from '@charmverse/core/prisma-client';
-import { currentSeason, getPreviousSeason, getSeasonWeekFromISOWeek } from '@packages/scoutgame/dates';
+import { currentSeason, getLastWeek, getPreviousSeason, getSeasonWeekFromISOWeek } from '@packages/scoutgame/dates';
 
 export type WeeklyReward = {
   week: string;
@@ -35,6 +35,7 @@ export async function getPointsWithEvents({
       recipientId: userId,
       claimedAt: isClaimed ? { not: null } : { equals: null },
       event: {
+        week: !isClaimed ? getLastWeek() : undefined,
         season: {
           in: claimableSeasons
         }
